requirements.txt
Babel==1.3
Flask>=0.8.0
Flask-Babel==0.9
Jinja2==2.7.1
PyRRD==0.1.0
Werkzeug==0.9.3
pudb==2013.3.5
python-rrdtool==1.4.7
pytz==2013d
coverage==3.6
nose==1.3.0
nosy==1.1.2
